requirements.txt
defusedxml==0.7.1
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.4
